+++ common/src/main/java/mca/client/gui/BlueprintScreen.java
@@ -146,6 +146,16 @@ public class BlueprintScreen extends ExtendedScreen {
         }
 
         switch (page) {
+            case "empty":
+                //add building
+                bx = width / 2 - 48;
+                by = height / 2;
+                addDrawableChild(new ButtonWidget(bx, by + 5, 96, 20, new TranslatableText("gui.blueprint.addBuilding"), (b) -> {
+                    NetworkHandler.sendToServer(new ReportBuildingMessage(ReportBuildingMessage.Action.ADD));
+                    NetworkHandler.sendToServer(new GetVillageRequest());
+                    close();
+                }));
+                break;
             case "refresh":
                 NetworkHandler.sendToServer(new ReportBuildingMessage(ReportBuildingMessage.Action.FULL_SCAN));
                 NetworkHandler.sendToServer(new GetVillageRequest());
@@ -310,7 +320,7 @@ public class BlueprintScreen extends ExtendedScreen {
                 drawCenteredText(transform, textRenderer, new TranslatableText("gui.blueprint.waiting"), width / 2, height / 2, 0xffaaaaaa);
                 break;
             case "empty":
-                drawCenteredText(transform, textRenderer, new TranslatableText("gui.blueprint.empty"), width / 2, height / 2, 0xffaaaaaa);
+                drawCenteredText(transform, textRenderer, new TranslatableText("gui.blueprint.empty"), width / 2, height / 2 - 20, 0xffaaaaaa);
                 break;
             case "map":
                 renderStats(transform);